Fixed test failures caused by reusing the same temp dir
